[tune] logger refactor part 1: move classes and utilities to own files
