+++ python/ray/tune/logger.py
@@ -1,12 +1,16 @@
 import csv
 import json
 import logging
-import os
-import yaml
 import numbers
+import os
+from typing import Callable, Dict, List, Optional, Type, Union
+
 import numpy as np
 
 import ray.cloudpickle as cloudpickle
+import yaml
+from ray.tune.trial import Trial
+from ray.tune.utils.util import SafeFallbackEncoder
 from ray.util.debug import log_once
 from ray.tune.result import (NODE_IP, TRAINING_ITERATION, TIME_TOTAL_S,
                              TIMESTEPS_TOTAL, EXPR_PARAM_FILE,
@@ -34,7 +38,10 @@ class Logger:
         trial (Trial): Trial object for the logger to access.
     """
 
-    def __init__(self, config, logdir, trial=None):
+    def __init__(self,
+                 config: Dict,
+                 logdir: str,
+                 trial: Optional[Trial] = None):
         self.config = config
         self.logdir = logdir
         self.trial = trial
@@ -89,7 +96,7 @@ class MLFLowLogger(Logger):
             client.log_param(self._run_id, key, value)
         self.client = client
 
-    def on_result(self, result):
+    def on_result(self, result: Dict):
         for key, value in result.items():
             if not isinstance(value, float):
                 continue
@@ -113,8 +120,8 @@ class JsonLogger(Logger):
         local_file = os.path.join(self.logdir, EXPR_RESULT_FILE)
         self.local_out = open(local_file, "a")
 
-    def on_result(self, result):
-        json.dump(result, self, cls=_SafeFallbackEncoder)
+    def on_result(self, result: Dict):
+        json.dump(result, self, cls=SafeFallbackEncoder)
         self.write("\n")
         self.local_out.flush()
 
@@ -127,7 +134,7 @@ class JsonLogger(Logger):
     def close(self):
         self.local_out.close()
 
-    def update_config(self, config):
+    def update_config(self, config: Dict):
         self.config = config
         config_out = os.path.join(self.logdir, EXPR_PARAM_FILE)
         with open(config_out, "w") as f:
@@ -136,7 +143,7 @@ class JsonLogger(Logger):
                 f,
                 indent=2,
                 sort_keys=True,
-                cls=_SafeFallbackEncoder)
+                cls=SafeFallbackEncoder)
         config_pkl = os.path.join(self.logdir, EXPR_PARAM_PICKLE_FILE)
         with open(config_pkl, "wb") as f:
             cloudpickle.dump(self.config, f)
@@ -159,7 +166,7 @@ class CSVLogger(Logger):
         self._file = open(progress_file, "a")
         self._csv_out = None
 
-    def on_result(self, result):
+    def on_result(self, result: Dict):
         tmp = result.copy()
         if "config" in tmp:
             del tmp["config"]
@@ -203,7 +210,7 @@ class TBXLogger(Logger):
         self._file_writer = SummaryWriter(self.logdir, flush_secs=30)
         self.last_result = None
 
-    def on_result(self, result):
+    def on_result(self, result: Dict):
         step = result.get(TIMESTEPS_TOTAL) or result[TRAINING_ITERATION]
 
         tmp = result.copy()
@@ -313,11 +320,11 @@ class UnifiedLogger(Logger):
     """
 
     def __init__(self,
-                 config,
-                 logdir,
-                 trial=None,
-                 loggers=None,
-                 sync_function=None):
+                 config: Dict,
+                 logdir: str,
+                 trial: Optional[Trial] = None,
+                 loggers: Optional[List[Type[Logger]]] = None,
+                 sync_function: Union[None, Callable, str] = None):
         if loggers is None:
             self._logger_cls_list = DEFAULT_LOGGERS
         else:
