+++ python/ray/tune/tests/test_function_api.py
@@ -9,7 +9,7 @@ from ray.rllib import _register_all
 
 from ray import tune
 from ray.tune.logger import NoopLogger
-from ray.tune.trainable import TrainableUtil
+from ray.tune.utils.trainable import TrainableUtil
 from ray.tune.function_runner import with_parameters, wrap_function, \
     FuncCheckpointUtil
 from ray.tune.result import TRAINING_ITERATION
