+++ python/ray/tune/utils/util.py
@@ -1,5 +1,7 @@
 import copy
+import json
 import logging
+import numbers
 import os
 import inspect
 import threading
@@ -11,6 +13,7 @@ from threading import Thread
 import numpy as np
 import ray
 import psutil
+import yaml
 
 logger = logging.getLogger(__name__)
 
@@ -536,6 +539,44 @@ def detect_config_single(func):
         logger.debug(str(e))
         use_config_single = False
     return use_config_single
+
+
+class SafeFallbackEncoder(json.JSONEncoder):
+    def __init__(self, nan_str="null", **kwargs):
+        super(SafeFallbackEncoder, self).__init__(**kwargs)
+        self.nan_str = nan_str
+
+    def default(self, value):
+        try:
+            if np.isnan(value):
+                return self.nan_str
+
+            if (type(value).__module__ == np.__name__
+                    and isinstance(value, np.ndarray)):
+                return value.tolist()
+
+            if issubclass(type(value), numbers.Integral):
+                return int(value)
+            if issubclass(type(value), numbers.Number):
+                return float(value)
+
+            return super(SafeFallbackEncoder, self).default(value)
+
+        except Exception:
+            return str(value)  # give up, just stringify it (ok for logs)
+
+
+def pretty_print(result):
+    result = result.copy()
+    result.update(config=None)  # drop config from pretty print
+    result.update(hist_stats=None)  # drop hist_stats from pretty print
+    out = {}
+    for k, v in result.items():
+        if v is not None:
+            out[k] = v
+
+    cleaned = json.dumps(out, cls=SafeFallbackEncoder)
+    return yaml.safe_dump(json.loads(cleaned), default_flow_style=False)
 
 
 if __name__ == "__main__":
